+++ src/Compze.DependencyInjection/Abstractions/_Compze.Utilities.DependencyInjection.Abstractions.cs
@@ -43,5 +43,5 @@ public enum Lifestyle
    Singleton,
    Scoped,
    TrackedTransient,
-   UntrackedTransient
+   Transient
 }